refactor extension tests
