+++ packages/fpdart/test/src/option_test.dart
@@ -164,34 +164,6 @@ void main() {
       });
     });
 
-    group('getOrElse', () {
-      test('Some', () {
-        final option = Option.of(10);
-        final value = option.getOrElse(() => 0);
-        expect(value, 10);
-      });
-
-      test('None', () {
-        final option = Option<int>.none();
-        final value = option.getOrElse(() => 0);
-        expect(value, 0);
-      });
-    });
-
-    group('alt', () {
-      test('Some', () {
-        final option = Option.of(10);
-        final value = option.alt(() => Option.of(0));
-        value.matchTestSome((some) => expect(some, 10));
-      });
-
-      test('None', () {
-        final option = Option<int>.none();
-        final value = option.alt(() => Option.of(0));
-        value.matchTestSome((some) => expect(some, 0));
-      });
-    });
-
     group('extend', () {
       test('Some', () {
         final option = Option.of(10);
@@ -507,15 +479,6 @@ void main() {
       final m2 = Option<int>.none();
       expect(m1.fold(() => 'none', (some) => 'some'), 'some');
       expect(m2.fold(() => 'none', (some) => 'some'), 'none');
-    });
-
-    test('elem', () {
-      final m1 = Option.of(10);
-      final m2 = Option<int>.none();
-      final eq = Eq.instance<int>((a1, a2) => a1 == a2);
-      expect(m1.elem(10, eq), true);
-      expect(m1.elem(9, eq), false);
-      expect(m2.elem(10, eq), false);
     });
 
     test('none()', () {
